refactor(stream): thin deepagents chat router
