+++ backend/tests/test_deepagents_shell.py
@@ -35,7 +35,7 @@ def _parse_sse_events(body: str) -> list[tuple[str, dict]]:
     return events
 
 
-def _patch_resolved_runtime(monkeypatch, dchat, captured: dict[str, object] | None = None):
+def _patch_resolved_runtime(monkeypatch, dstream, captured: dict[str, object] | None = None):
     def _fake_resolve_deepagents_runtime(
         db,
         *,
@@ -70,7 +70,7 @@ def _patch_resolved_runtime(monkeypatch, dchat, captured: dict[str, object] | No
             limiter=SimpleNamespace(),
         )
 
-    monkeypatch.setattr(dchat, "resolve_deepagents_runtime", _fake_resolve_deepagents_runtime)
+    monkeypatch.setattr(dstream, "resolve_deepagents_runtime", _fake_resolve_deepagents_runtime)
 
 
 @pytest.mark.integration
@@ -80,14 +80,14 @@ def test_deepagents_chat_stream_basic(monkeypatch):
     client = _create_test_client()
     headers = _auth_headers(client)
 
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
     from app.services.deepagents import deepagents_graph as dag
     from app.services.deepagents.tool_runtime import ToolPolicyUsage
 
-    _patch_resolved_runtime(monkeypatch, dchat)
+    _patch_resolved_runtime(monkeypatch, dstream)
     monkeypatch.setattr(
-        dchat,
-        "_serialize_agent_topology",
+        dstream,
+        "serialize_agent_topology",
         lambda agent: {
             "nodes": [
                 {"id": "__start__", "name": "__start__", "kind": "start"},
@@ -162,7 +162,7 @@ def test_deepagents_chat_stream_basic(monkeypatch):
         return _FakeAgent(), ToolPolicyUsage(), [], {}
 
     monkeypatch.setattr(dag, "build_chat_agent", _fake_build_chat_agent)
-    monkeypatch.setattr(dchat, "build_chat_agent", _fake_build_chat_agent)
+    monkeypatch.setattr(dstream, "build_chat_agent", _fake_build_chat_agent)
 
     with client.stream(
         "POST",
@@ -229,11 +229,11 @@ def test_deepagents_chat_stream_accepts_fetch_transport_payload(monkeypatch):
     client = _create_test_client()
     headers = _auth_headers(client)
 
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
     from app.services.deepagents import deepagents_graph as dag
     from app.services.deepagents.tool_runtime import ToolPolicyUsage
 
-    _patch_resolved_runtime(monkeypatch, dchat)
+    _patch_resolved_runtime(monkeypatch, dstream)
 
     captured: dict[str, object] = {}
 
@@ -257,7 +257,7 @@ def test_deepagents_chat_stream_accepts_fetch_transport_payload(monkeypatch):
         return _FakeAgent(), ToolPolicyUsage(), [], {}
 
     monkeypatch.setattr(dag, "build_chat_agent", _fake_build_chat_agent)
-    monkeypatch.setattr(dchat, "build_chat_agent", _fake_build_chat_agent)
+    monkeypatch.setattr(dstream, "build_chat_agent", _fake_build_chat_agent)
 
     with client.stream(
         "POST",
@@ -367,7 +367,7 @@ def test_deepagents_chat_stream_worker_uses_owned_session(monkeypatch):
     client = _create_test_client()
     headers = _auth_headers(client)
 
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
     from app.services.deepagents.tool_runtime import ToolPolicyUsage
 
     worker_session = SimpleNamespace(closed=False)
@@ -378,9 +378,9 @@ def test_deepagents_chat_stream_worker_uses_owned_session(monkeypatch):
 
     worker_session.close = _fake_close
 
-    monkeypatch.setattr(dchat, "create_session", lambda: worker_session)
-    _patch_resolved_runtime(monkeypatch, dchat, captured)
-    monkeypatch.setattr(dchat, "_serialize_agent_topology", lambda agent: None)
+    monkeypatch.setattr(dstream, "create_session", lambda: worker_session)
+    _patch_resolved_runtime(monkeypatch, dstream, captured)
+    monkeypatch.setattr(dstream, "serialize_agent_topology", lambda agent: None)
 
     class _FakeAgent:
         def stream(self, payload, **kwargs):  # noqa: ANN001
@@ -392,7 +392,7 @@ def test_deepagents_chat_stream_worker_uses_owned_session(monkeypatch):
             }
 
     monkeypatch.setattr(
-        dchat,
+        dstream,
         "build_chat_agent",
         lambda **kwargs: (_FakeAgent(), ToolPolicyUsage(), [], {}),
     )
@@ -456,12 +456,12 @@ def test_deepagents_chat_stream_filters_draft_tool_calls_without_tool_run_custom
     client = _create_test_client()
     headers = _auth_headers(client)
 
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
     from app.services.deepagents import deepagents_graph as dag
     from app.services.deepagents.tool_runtime import ToolPolicyUsage
 
-    _patch_resolved_runtime(monkeypatch, dchat)
-    monkeypatch.setattr(dchat, "_serialize_agent_topology", lambda agent: None)
+    _patch_resolved_runtime(monkeypatch, dstream)
+    monkeypatch.setattr(dstream, "serialize_agent_topology", lambda agent: None)
 
     class _FakeAgent:
         def stream(self, payload, **kwargs):  # noqa: ANN001
@@ -493,7 +493,7 @@ def test_deepagents_chat_stream_filters_draft_tool_calls_without_tool_run_custom
         return _FakeAgent(), ToolPolicyUsage(), [], {}
 
     monkeypatch.setattr(dag, "build_chat_agent", _fake_build_chat_agent)
-    monkeypatch.setattr(dchat, "build_chat_agent", _fake_build_chat_agent)
+    monkeypatch.setattr(dstream, "build_chat_agent", _fake_build_chat_agent)
 
     with client.stream(
         "POST",
@@ -515,15 +515,15 @@ def test_deepagents_chat_stream_filters_draft_tool_calls_without_tool_run_custom
 
 
 def test_sanitize_tool_call_drops_empty_json_string_args():
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
 
-    assert dchat._sanitize_tool_call(
+    assert dstream._sanitize_tool_call(
         {"id": "call-1", "name": "web_search", "args": "{}"}
     ) is None
-    assert dchat._sanitize_tool_call(
+    assert dstream._sanitize_tool_call(
         {"id": "call-1", "name": "web_search", "args": "[]"}
     ) is None
-    assert dchat._sanitize_tool_call(
+    assert dstream._sanitize_tool_call(
         {"id": "call-1", "name": "web_search", "args": "null"}
     ) is None
 
@@ -645,13 +645,13 @@ def test_deepagents_chat_stream_emits_run_timeout_and_done(monkeypatch):
     client = _create_test_client()
     headers = _auth_headers(client)
 
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
     from app.services.deepagents import deepagents_graph as dag
     from app.services.deepagents.tool_runtime import ToolPolicyUsage
 
-    _patch_resolved_runtime(monkeypatch, dchat)
-    monkeypatch.setattr(dchat, "_serialize_agent_topology", lambda agent: None)
-    monkeypatch.setattr(dchat.settings, "deepagents_run_timeout_seconds", 0.1, raising=False)
+    _patch_resolved_runtime(monkeypatch, dstream)
+    monkeypatch.setattr(dstream, "serialize_agent_topology", lambda agent: None)
+    monkeypatch.setattr(dstream.settings, "deepagents_run_timeout_seconds", 0.1, raising=False)
 
     class _FakeAgent:
         def stream(self, payload, **kwargs):  # noqa: ANN001
@@ -665,7 +665,7 @@ def test_deepagents_chat_stream_emits_run_timeout_and_done(monkeypatch):
         return _FakeAgent(), ToolPolicyUsage(), [], {}
 
     monkeypatch.setattr(dag, "build_chat_agent", _fake_build_chat_agent)
-    monkeypatch.setattr(dchat, "build_chat_agent", _fake_build_chat_agent)
+    monkeypatch.setattr(dstream, "build_chat_agent", _fake_build_chat_agent)
 
     with client.stream(
         "POST",
@@ -689,7 +689,7 @@ def test_deepagents_chat_stream_client_disconnect_cancels_worker_and_closes_sess
     client = _create_test_client()
     headers = _auth_headers(client)
 
-    import app.routers.deepagents_chat as dchat
+    import app.services.deepagents.stream_gateway as dstream
     from app.services.deepagents.tool_runtime import ToolPolicyUsage
 
     cancel_seen = threading.Event()
@@ -702,9 +702,9 @@ def test_deepagents_chat_stream_client_disconnect_cancels_worker_and_closes_sess
 
     worker_session.close = _fake_close
 
-    monkeypatch.setattr(dchat, "create_session", lambda: worker_session)
-    _patch_resolved_runtime(monkeypatch, dchat, captured)
-    monkeypatch.setattr(dchat, "_serialize_agent_topology", lambda agent: None)
+    monkeypatch.setattr(dstream, "create_session", lambda: worker_session)
+    _patch_resolved_runtime(monkeypatch, dstream, captured)
+    monkeypatch.setattr(dstream, "serialize_agent_topology", lambda agent: None)
 
     class _FakeAgent:
         def stream(self, payload, **kwargs):  # noqa: ANN001
@@ -726,7 +726,7 @@ def test_deepagents_chat_stream_client_disconnect_cancels_worker_and_closes_sess
         captured["cancel_token"] = kwargs["cancel_token"]
         return _FakeAgent(), ToolPolicyUsage(), [], {}
 
-    monkeypatch.setattr(dchat, "build_chat_agent", _fake_build_chat_agent)
+    monkeypatch.setattr(dstream, "build_chat_agent", _fake_build_chat_agent)
 
     with client.stream(
         "POST",
